requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
datasets==3.6.0
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl#sha256=fec69fec52b1780f2d269d5af7582a5e28028738bd3190532459aeb473bfa3e7
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
huggingface-hub==0.31.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.4.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
wrapt==1.17.2
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.0
